템플릿메소드
