+++ java.study/src/main/java/designpattern/creation/templatemethod/ex2/TemplateMethodPattern.java
@@ -1,4 +1,4 @@
-package designpattern.templatemethod.ex2;
+package designpattern.creation.templatemethod.ex2;
 
 /**
  * <pre>
