--- java.study/src/main/java/designpattern/templatemethod/ex2/Genesis.java
@@ -1,15 +1,0 @@
-package designpattern.templatemethod.ex2;
-
-/**
- * <pre>
- * Description :
- *
- *
- * </pre>
- *
- * @author skan
- * @version Copyright (C) 2022 by CJENM|MezzoMedia. All right reserved.
- * @since 2022/11/30
- */
-public class Genesis extends HyundaiCar {
-}
